♻️ refactor(bot): extract startup helpers into bot/utils/helpers
Move `_run_webhook` and `notify_admins_on_startup` out of
`presentation/bot/main.py` into a dedicated `presentation/bot/utils/helpers.py`
module so `main.py` focuses on wiring (DI, routers, mode dispatch) and the
supporting coroutines live next to the rest of the bot utilities.
